Make P[] static
P[] and M are the only variable that is shared across the two codes.
Once P[] is made static, it is easy to split the two functionalities into
separate methods. Once again, it is not a good style to create static
attributes. In later commits, I will make this an object attribute.